+++ src/main/java/org/zerock/myapp/mapper/AttachMapper.java
@@ -9,8 +9,6 @@ public interface AttachMapper {
 	/* 이미지 데이터 반환 */
 	public List<AttachImageVO> getAttachList(int bookId);
 	
-	// test 주석입니당.
-	// test 주석2
-	//ㅋㅋ 메롱
+
 
 } // end class
