+++ src/main/resources/org/zerock/myapp/mapper/MemberMapper.xml
@@ -6,27 +6,21 @@
 
 <mapper namespace="org.zerock.myapp.mapper.MemberMapper">
 
-	<!-- 성별, 회원가입날짜, 생년월일 insert문 돌아가게 설정해줘야함!! -->
-    <!-- 회원가입: memberJoin -->
- <!--     <insert id="memberJoin" parameterType="org.zerock.myapp.domain.MemberVO">
-        INSERT INTO MEMBER_TEST
-        VALUES(#{MEMBER_ID}, #{MEMBER_PW}, #{MEMBER_NAME}, #{MEMBER_MAIL}, #{MEMBER_HP}, #{MEMBER_ADDR1}, #{MEMBER_ADDR2}, #{MEMBER_ZIPCODE}, 0)
-    </insert> -->
 
      <!-- 로그인: memberLogin -->
      <select id="memberLogin" resultType="org.zerock.myapp.domain.UserVO">
          SELECT MEMBER_ID, MEMBER_NAME, ADMINCK FROM MEMBER_TEST WHERE MEMBER_ID = #{MEMBER_ID} and MEMBER_PW = #{MEMBER_PW}
      </select>
      
-     <!-- 
-     <select id="selectUser" resultType="org.zerock.myapp.domain.UserVO">
-		SELECT * 
-		FROM tbl_user 
-		WHERE 
-			userid = #{userid} 
-			AND userpw = #{userpw}
-	</select>
-	-->
+	     <!-- 
+	     <select id="selectUser" resultType="org.zerock.myapp.domain.UserVO">
+			SELECT * 
+			FROM tbl_user 
+			WHERE 
+				userid = #{userid} 
+				AND userpw = #{userpw}
+		</select>
+		-->
 	<!-- 혁규 회원가입 -->
 	  <insert id="memberSignup" >
   INSERT INTO member (
